Refactor sendPFSTransactionEmail deploys to AWS
https://github.com/SCPR/digital_products/issues/127

The lambda function can be deployed to the web using serverless. The default deployment mode is 'dev', which will email louise.yang@scpr.org and not set the `emailSent` flag to true.  This way, developers don't have to keep reverting the flag when testing.

To deploy the production version, use `-s prod` when running the `serverless deploy` method.